requirements.txt
cryptography>=41.0.0
customtkinter>=5.2.0
argon2-cffi>=23.0.0
portalocker>=2.7.0
Pillow>=10.0.0
